Update: comment

+++ twitkit-fridge/twitkit-fridge-src/src/main/java/com/enkanrec/twitkitFridge/api/form/CommentForm.java
@@ -23,7 +23,7 @@ public class CommentForm extends JsonDataFridgeForm {
      * 任务id
      */
     @NotNull
-    private Long tid;
+    private Integer tid;
 
     /**
      * 备注内容
